requirements.txt
Django==1.8.1
django-nose
django-autoslug
django-taggit
loremipsum
elasticsearch
django-haystack
-e git://github.com/lfalvarez/django-candidator.git#egg=django-candidator
PyYAML
django-markdown-deux==1.0.4
django-extensions
sorl-thumbnail
pillow
pgmagick
django_admin_bootstrapped
django-tinymce
django-registration
mimeparse
django-tastypie
django-pagination==1.0.5
django-celery
python-memcached
-e git+https://github.com/openpolis/django-popolo.git#egg=django_popolo-dev
django-mathfilters
django-picklefield
tropo-webapi-python